Refactoring to Multiple Files
Actually I'm starting to find wave equations and others quite hard to grasp, I might be wrong on some parts and may need some rework.
