+++ crates/energy/waves/src/lib.rs
@@ -1,2 +1,4 @@
+pub mod oscillation;
 pub mod propagation;
+pub mod superposition;
 pub mod wave_equation;
